Implement Mudcrack Y game
Adapt the existing Y implementation to play mudcrack Y instead.
The board is represented using a graph of the cells, and
connectivity is tracked using union-find. Since this is mudcrack Y
the board graph is currently written in by hand, and right now defaults
to a base 3 version of geodesic Y.